run zend without tracker

+++ zend/entrypoint.sh
@@ -52,8 +52,8 @@ fi
 
 # Copy in any additional SSL trusted CA
 if [ -d "/mnt/zen/certs" ]; then
-  domain="$(cat /mnt/zen/secnode/fqdn)"
-  if [ -f /mnt/zen/certs/$domain/ca.cer ]; then
+  domain="$Z_DOMAIN"
+  if [ ! -z $domain ] && [ -f /mnt/zen/certs/$domain/ca.cer ]; then
     echo "Copying additional trusted SSL certificates"
     cp /mnt/zen/certs/$domain/ca.cer /usr/local/share/ca-certificates/ca.crt > /dev/null 2>&1 || true
     update-ca-certificates --fresh
@@ -68,6 +68,6 @@ chown -R user:user /home/user /mnt/zen
 echo "Starting $@ .."
 if [[ "$1" == zend ]]; then
     exec /usr/local/bin/gosu user /bin/bash -c "$@ $OPTS"
+else
+    exec /usr/local/bin/gosu user "$@"
 fi
-
-exec /usr/local/bin/gosu user "$@"
